chore: refactor for modularity
