+++ src/components/Home/GroupDetails/index.css
