+++ src/components/JobBoard/index.js
@@ -1,0 +1,4 @@
+import React from 'react'
+import './index.css'
+
+export default props => <div>This is the job board page</div>
