--- src/components/NavBar.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import {Link} from 'react-router-dom'
-import '../styles/NavBar.css'
-
-const NavBar = props => (
-    <ul className='NavBar'>
-        <li>Home</li>
-        <li>Contact</li>
-    </ul>
-)
-
-export default NavBar
