+++ src/components/NotFound/index.js
@@ -1,0 +1,4 @@
+import React from 'react'
+import './index.css'
+
+export default props => <div>404 page not found</div>
