+++ src/shared/Footer/index.js
@@ -1,0 +1,4 @@
+import React from 'react'
+import './index.css'
+
+export default props => <div>This will show up as the footer on every page</div>
