+++ src/shared/SocialMediaIcons/index.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import {SocialIcons} from 'react-social-icons';
 
 //styles
-import '../styles/SocialMediaIcons.css'
+import './index.css'
 
 const SocialMediaIcons = props => (
     <div className='SocialMediaIcons'>
